requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
aiokafka==0.10.0
redis==5.0.1
pyspark==3.5.0
kafka-python==2.0.2
boto3==1.34.34
pandas==2.2.0
pyarrow==14.0.2
psycopg2-binary==2.9.9
openai==1.12.0
httpx==0.27.2
psycopg2-binary==2.9.9
